feat: switch to i32

+++ src/dto/animal_response_dto.rs
@@ -2,10 +2,10 @@ use serde::{Serialize};
 
 #[derive(Serialize)]
 pub struct AnimalResponseDto {
-    pub id: i64,
+    pub id: i32,
     pub species: String,
     pub common_name: String,
     pub habitat: String,
-    pub lifespan: i64,
+    pub lifespan: i32,
     pub is_endangered: bool
 }
